includes WC_VICTORIABANK_PLUGIN_FILE
